discord-bot/requirements.txt
discord.py
chromadb
pysqlite3-binary
uuid
nextcord
python-dotenv